refactor: move provider and selector components to components/result folder
EX-6762
